+++ tests/testthat/test-labelled.r
@@ -790,7 +790,7 @@ test_that("set_value_labels replaces all value labels", {
   expect_null(val_labels(v))
 })
 
-test_that("set_value_labels .labels + .overwrite = FALSE on data frames", {
+test_that("set_value_labels .labels + .overwrite on data frames", {
   df <- data.frame(
     s1 = labelled(c("M", "M", "F"), c(Male = "M", Female = "F")),
     s2 = labelled(c(1, 1, 2), c(Yes = 1, No = 2)),
@@ -807,6 +807,16 @@ test_that("set_value_labels .labels + .overwrite = FALSE on data frames", {
   )
   expect_equal(val_labels(df2$s1), c(Male = "M", Female = "F"))
   expect_equal(val_labels(df2$s2), c(Yes = 1, No = 2))
+  expect_equal(val_labels(df2$x), c(low = 1))
+
+  # .overwrite = TRUE
+  df2 <- set_value_labels(
+    df,
+    .labels = list(s1 = c(Male2 = "M"), s2 = c(Maybe = 3), x = c(low = 1)),
+    .overwrite = TRUE
+  )
+  expect_equal(val_labels(df2$s1), c(Male2 = "M"))
+  expect_equal(val_labels(df2$s2), c(Maybe = 3))
   expect_equal(val_labels(df2$x), c(low = 1))
 })