+++ NEWS.md
@@ -6,6 +6,7 @@
   only `NA` values (#192)
 * `dictionary_to_value_labels()` now handles value labels with repeated
   `delim_value_label` (#194)
+* fix in `set_value_labels()` when `.overwrite = FALSE` (#198, @ LeonidasZhak)
 
 # labelled 2.16.0